% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoders.R
\name{encoder_date}
\alias{encoder_date}
\title{Date Encoder}
\usage{
encoder_date(column, name = column, time_of_day = NULL,
  weekend = NULL)
}
\arguments{
\item{column}{Column name to use as input.}

\item{name}{The name for this encoder, defaults to \code{column}.}

\item{time_of_day}{TODO}

\item{weekend}{TODO}
}
\description{
Date Encoder
}
\seealso{
\code{\link{nupic_encoders}}.
}
