% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoders.R
\name{encoder_random_distributed_scalar}
\alias{encoder_random_distributed_scalar}
\title{Random Distributed Scalar Encoder}
\usage{
encoder_random_distributed_scalar(column, name = column, resolution)
}
\arguments{
\item{column}{Column name to use as input.}

\item{name}{The name for this encoder, defaults to \code{column}.}

\item{resolution}{TODO}
}
\description{
Random Distributed Scalar Encoder
}
\seealso{
\code{\link{nupic_encoders}}.
}
