% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoders.R
\name{encoder_sdr_category}
\alias{encoder_sdr_category}
\title{SDR Category Encoder}
\usage{
encoder_sdr_category(column, name = column, n, w, category_list = NULL,
  verbosity = 0, seed = 1)
}
\arguments{
\item{column}{Column name to use as input.}

\item{name}{The name for this encoder, defaults to \code{column}.}

\item{n}{The number of bits to use.}

\item{w}{The width of the SDR.}

\item{category_list}{List of discrete string categories, if \code{NULL} 
then categories will be added automatically as they are encountered.}

\item{verbosity}{An integer controlling the level of debugging output. A
value of zero implies no output.}

\item{seed}{The seed used for numpy's random number generator.}
}
\description{
Encodes a list of discrete categories (described by strings), that aren't
related to each other. Each encoding is an SDR in which \code{w} out of
\code{n} bits are turned on.
}
