% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gym_hourly}
\alias{gym_hourly}
\title{Gym Hourly Data}
\format{An object of class \code{"data.frame"}.}
\source{
\href{https://github.com/numenta/nupic}{Nupic}
}
\usage{
data(gym_hourly)
}
\description{
Gym Hourly Data
}
\examples{
data(gym_hourly)
}
\keyword{datasets}
\keyword{internal}
