% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{nupic}
\alias{nupic}
\title{Nupic Model}
\usage{
nupic(data = gym_hourly[1:50, ], predict = "consumption",
  encoders = nupic_config_en(encoder_random_distributed_scalar("consumption",
  resolution = 0.88, seed = 1), encoder_date("timestamp", "time_of_day",
  time_of_day = list(21L, 1L)), encoder_date("timestamp", "weekend",
  weekend = 21L)), tm = nupic_config_tm(), sp = nupic_config_sp(),
  cl = nupic_config_cl(), se = nupic_confif_se())
}
\arguments{
\item{data}{The data frame to use for online prediction.}

\item{predict}{The column name in \code{data} to predict from.}

\item{encoders}{The sensor encodings for the input columns using
\code{\link{nupic_config_en}}.}

\item{tm}{The TM configuration, use \code{nupic_config_tm()}.}

\item{sp}{The SP configuration, use \code{nupic_config_sp()}.}

\item{se}{The SE configuration, use \code{nupic_config_se()}.}
}
\description{
Runs a Nupic model for online predictions using the cortical
learning algorithm.
}
