% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{nupic_confif_se}
\alias{nupic_confif_se}
\title{Sensors Configuration}
\usage{
nupic_confif_se(verbosity = 0, sensor_auto_reset = NULL)
}
\arguments{
\item{verbosity}{Diagnostic output verbosity control. Zero as silent,
\code{1:6} increasing levels of verbosity.}

\item{sensor_auto_reset}{A list specifying the period for
automatically-generated resets from a RecordSensor. \code{NULL}
disable automatically-generated resets (also disabled if 
all of the specified values evaluate to 0). Valid keys is the desired
combination of the following: days, hours, minutes, seconds,
milliseconds, microseconds, weeks.}
}
\description{
Sensors Configuration
}
\examples{

nupic_config_se(sensor_auto_reset = list(days = 1L, hours = 12L))
  
}
