% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{nupic_config_cl}
\alias{nupic_config_cl}
\title{Cortical Learning Configuration}
\usage{
nupic_config_cl(verbosity = 0, alpha = 0.1, steps = c(1, 5))
}
\arguments{
\item{verbosity}{Diagnostic output verbosity control. Zero as silent,
\code{1:6} increasing levels of verbosity.}

\item{alpha}{This controls how fast the classifier learns/forgets. Higher
values make it adapt faster and forget older patterns faster.}

\item{steps}{This is set after the call to updateConfigFromSubConfig and is
computed from the aggregationInfo and predictAheadTime.}
}
\description{
Cortical Learning Configuration
}
