% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{nupic_config_en}
\alias{nupic_config_en}
\title{Encoders Configuration}
\usage{
nupic_config_en(...)
}
\arguments{
\item{...}{One or many encoders.}
}
\description{
Encoders Configuration
}
\seealso{
\code{\link{encoder_random_distributed_scalar}},
         \code{\link{encoder_date}}.
}
