% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{nupic_config_sp}
\alias{nupic_config_sp}
\title{Spatial Pooler Configuration}
\usage{
nupic_config_sp(verbosity = 0, global_inhibition = 1,
  column_count = 2048, input_width = 0,
  num_active_columns_in_area = 40, seed = 1956, potential_pct = 0.85,
  syn_perm_connected = 0.1, syn_perm_active_inc = 0.04,
  syn_perm_inactive_dec = 0.005, boost_strength = 3)
}
\arguments{
\item{verbosity}{SP diagnostic output verbosity control,
zero for silent.}

\item{global_inhibition}{Global inhibition.}

\item{column_count}{Number of cell columns in the cortical
region (same number for SP and TM).}

\item{input_width}{Input width.}

\item{num_active_columns_in_area}{SP inhibition control (absolute value);
Maximum number of active columns in the SP region's output (when there
are more, the weaker ones are suppressed).}

\item{seed}{Integer number to use as seed.}

\item{potential_pct}{What percent of the columns's receptive field is
available for potential synapses.}

\item{syn_perm_connected}{The default connected threshold. Any synapse
whose permanence value is above the connected threshold is a
"connected synapse", meaning it can contribute to the cell's firing.
Typical value is 0.10.}

\item{syn_perm_active_inc}{Active synapse permanence increment.}

\item{syn_perm_inactive_dec}{Inactive synapse permanence decrement.}

\item{boost_strength}{boostStrength controls the strength of boosting.
It should be a number greater or equal than 0.0. No boosting is applied
if it's zero, boosting encourages efficient usage of SP columns.}
}
\description{
Spatial Pooler Configuration
}
