% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{nupic_config_tm}
\alias{nupic_config_tm}
\title{Temporal Model Configuration}
\usage{
nupic_config_tm(verbosity = 0, column_count = 2048,
  cells_per_column = 32, input_width = 2048, seed = 1960,
  new_synapse_count = 20, max_synapses_per_segment = 32,
  max_segments_per_cell = 128, initial_perm = 0.21,
  permanence_inc = 0.1, permanence_dec = 0.1, global_decay = 0,
  max_age = 0, min_threshold = 12, activation_threshold = 16,
  pam_length = 1)
}
\arguments{
\item{verbosity}{Diagnostic output verbosity control. Zero as silent,
\code{1:6} increasing levels of verbosity.}

\item{column_count}{Number of cell columns in the cortical region
(same number for SP and TM).}

\item{cells_per_column}{The number of cells (i.e., states), allocated
per column.}

\item{input_width}{The width of the input.}

\item{seed}{Integer number to use as seed.}

\item{new_synapse_count}{New Synapse formation count. The number of
synapses added to a segment during learning}

\item{max_synapses_per_segment}{Maximum number of synapses per segment.
Bigger than zero for fixed-size CLA and \code{-1} for non-fixed-size
CLA.}

\item{max_segments_per_cell}{Maximum number of segments per cell.
Biggeer than zero for fixed-size CLA and \code{-1} for non-fixed-size
CLA.}

\item{initial_perm}{Initial permanence for newly created synapses.}

\item{permanence_inc}{Permanence Increment.}

\item{permanence_dec}{Permanence Decrement. If set to \code{NULL}, will
automatically default to 'tpPermanenceIncvalue'.}

\item{global_decay}{Global decay.}

\item{max_age}{Max age.}

\item{min_threshold}{Minimum number of active synapses for a segment
to be considered during search for the best-matching segments. If
set to \code{NULL} it uses the default, replaces 'tpMinThreshold'.}

\item{activation_threshold}{Segment activation threshold. A segment is
active if it has more or equal 'tpSegmentActivationThreshold' connected
synapses that are active due to infActiveState. If set to \code{NULL}
it uses the default, replaces 'tpActivationThreshold'.}

\item{pam_length}{"Pay Attention Mode" length. This tells the TM how
many new elements to append to the end of a learned sequence at a time.
Smaller values are better for datasets with short sequences, higher
values are better for datasets with long sequences.}
}
\description{
Temporal Model Configuration
}
