% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoders.R
\name{nupic_encoders}
\alias{nupic_encoders}
\title{Define Encoders}
\usage{
nupic_encoders(...)
}
\description{
Define Encoders
}
\seealso{
\code{\link{encoder_random_distributed_scalar}},
         \code{\link{encoder_date}}.
}
