% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nyc_taxi}
\alias{nyc_taxi}
\title{New York Taxis}
\format{An object of class \code{"data.frame"}.}
\source{
\href{https://github.com/numenta/nupic}{Nupic}
}
\usage{
data(nyc_taxi)
}
\description{
New York Taxis
}
\examples{
data(nyc_taxi)
}
\keyword{datasets}
\keyword{internal}
